Add files via upload
Dyslexia-Friendly Web App is designed to create an accessible and user-friendly experience for individuals with dyslexia. Built with Django, this app offers a range of features to enhance readability and usability. Key features include customizable font options, text-to-speech functionality, and adjustable color themes to cater to different preferences and needs.